* develop:
  - resolved couple of warnings
  - remove commented code
  - raised app version to 1.0(31)
  - fixed issue with crashing after calling side menu on About project screen
  - update password and email warning for iPad in login screen
  - changed cell type for RolesViewController
  - fixed issue with logout - fixed issue with crashing on showing image picker controller for access to the photo library or camera
  - fixed issue with reloading data after updating team member role info
  - fixed back button for iPad in TaskDetail VC - fixed Cancel button action for invite team member VC
  - fixed phone button for team members
  - resolved issue with showing side menu
  - fixed issue with selection role on teamProfileController
  - fixed ui bug in tasks controller
  - added highlighting selected project cell in main menu for iPhone.
  - added fix constraint for team info - added fix constrain for alert

+++ src/Helpers/Utils.m
@@ -439,7 +439,7 @@ static NSDateFormatter* defaultDateFormatter = nil;
 
 + (NSString*) generateStringOfDaysCount: (NSUInteger) count
 {
-    NSString* daysCountString = [NSString stringWithFormat: @"%ld ", count];
+    NSString* daysCountString = [NSString stringWithFormat: @"%ld ", (unsigned long)count];
     NSString* sufixString     = @"";
     
     switch (count)
